Rename `commands` -> `updates`
