+++ identity-account/src/tests/mod.rs
@@ -2,5 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 mod account;
-mod commands;
 mod lazy;
+mod updates;
